engine: remove util dependency
